requirements.txt
accelerate==1.6.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
diffusers==0.33.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.54
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.32
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
ollama==0.4.8
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pillow==11.2.1
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.3
tenacity==9.1.2
tokenizers==0.21.1
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.2
zipp==3.21.0
zstandard==0.23.0
